Move TransactionFilterTypes to the Models/ directory
